requirements.txt
beautifulsoup4>=4.9.1
click>=7.1.2
cssselect>=1.1.0
date-extractor>=4.1.1
fake_useragent>=0.1.11
feedfinder2>=0.0.4
feedparser>=5.2.1
google-cloud-translate>=2.0.1
jieba3k>=0.35.1
lxml>=4.5.2
numpy>=1.19.0
pandas>=1.0.5
pdftotext>=2.1.4
Pillow>=7.2.0
pyarabic>=0.6.8
python-dotenv>=0.14.0
pytest>=5.4.3
pytextrank>=2.0.2
pythainlp>=2.2.2
python-dateutil>=2.8.1
dateparser>=0.7.6
PyPDF4>=1.27.0
PyYAML>=5.3.1
requests>=2.24.0
requests_toolbelt>=0.9.1
spacy>=2.3.1
tldextract>=2.2.2
tinysegmenter==0.4
flask>=1.1.2
flask-cors>=3.0.8
flask_restful>=0.3.8
google-api-python-client>=1.9.3
gunicorn>=20.0.4
oauth2-client>=1.2.1
pytest-flask>=1.0.0
waitress>=1.4.4
